pyproject.toml
[tool.poetry]
name = "invesnote"
version = "0.1.0"
description = ""
authors = ["Daniel <dg1638382@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.33.1"
beautifulsoup4 = "^4.14.3"
tavily-python = "^0.7.23"
python-dotenv = "^1.2.2"
fastapi = "^0.136.1"
uvicorn = "^0.46.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
